requirements/requirements.txt
Django>=3.0
django-cors-headers>=3.5.0
django-extensions>=3.0.5
django-filter>=2.3.0
djangorestframework>=3.11.1
shortuuid>=1.0.1
psycopg2-binary>=2.8
-r dev-requirements.txt